Moving ContentCardLink types to new file
